requirements.txt
async-timeout>=4.0.2
blurhash>=1.1.4
certifi>=2022.6.15
charset-normalizer>=2.0.12
decorator>=5.1.1
Deprecated>=1.2.13
idna>=3.3
Mastodon.py>=1.5.1
oauthlib>=3.2.0
packaging>=21.3
pyparsing>=3.0.9
PySocks>=1.7.1
python-dateutil>=2.8.2
python-dotenv>=0.20.0
python-magic>=0.4.27
pytz>=2022.1
redis>=4.3.3
requests>=2.28.0
requests-oauthlib>=1.3.1
six>=1.16.0
tweepy>=4.10.0
urllib3>=1.26.9
wrapt>=1.14.1
